각종 수치 조정과 경험치 구현중

+++ Neo/Shadow In The Shade/Assets/01.Scripts/Enemy/Slime/Slime_Bone.cs
@@ -15,7 +15,7 @@ public class Slime_Bone : Enemy, ITacklable
     private Move_Chase chase = null;
     private Attack_Tackle attack = null;
 
-    int defaultDamage;
+    float defaultDamage;
     readonly int damageIncreaseAmout = 2;
 
     private readonly WaitForSeconds halfSecWait = new WaitForSeconds(0.5f);
